figure: Initial import
This marks the beginning of Commishes as a single repository, allowing
me to version all the components at once. This makes this repository the
single source of truth.

The intent here is to reduce the overhead of maintaining dozens of
individual repos for each component and instead using tree splitting to
export the components to repos.